+++ target-stable/doc/crates.js
@@ -1,2 +1,0 @@
-window.ALL_CRATES = ["tf_tree"];
-//{"start":21,"fragment_lengths":[9]}