+++ target-stable/doc/src-files.js
@@ -1,2 +1,0 @@
-createSrcSidebar('[["tf_tree",["",[],["cache.rs","lib.rs","tree.rs"]]]]');
-//{"start":19,"fragment_lengths":[51]}