+++ target-stable/doc/trait.impl/core/default/trait.Default.js
@@ -1,9 +1,0 @@
-(function() {
-    const implementors = Object.fromEntries([["tf_tree",[["impl <a class=\"trait\" href=\"https://doc.rust-lang.org/1.97.1/core/default/trait.Default.html\" title=\"trait core::default::Default\">Default</a> for <a class=\"struct\" href=\"tf_tree/struct.TreeBuilder.html\" title=\"struct tf_tree::TreeBuilder\">TreeBuilder</a>",0]]]]);
-    if (window.register_implementors) {
-        window.register_implementors(implementors);
-    } else {
-        window.pending_implementors = implementors;
-    }
-})()
-//{"start":59,"fragment_lengths":[287]}